fix: replaces std::powl with std::pow since there is an overload
- std::powl() is only available in gcc-14 see
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=79700
- std::pow() overload does the same, didn't notice the overload at first

+++ include/raytracing_scene.hpp
@@ -121,12 +121,12 @@ class RaytracingScene
 
 			std::vector<BezierTriangle2> subTriangles;
 			subTriangles.reserve(
-			    static_cast<size_t>(std::powl(4.0L, static_cast<long double>(subdivisions))));
+			    static_cast<size_t>(std::pow(4.0L, static_cast<long double>(subdivisions))));
 			subTriangles.push_back(bezierTriangle);
 
 			for (int i = 0; i < subdivisions; i++)
 			{
-				for (size_t j = 0; j < std::powl(4.0L, static_cast<long double>(i)); j++)
+				for (size_t j = 0; j < std::pow(4.0L, static_cast<long double>(i)); j++)
 				{
 					BezierTriangle2& t = subTriangles.front();
 					const auto& s = ltracer::subdivideBezierTriangle2(t);
